requirements.txt
python-telegram-bot==21.9
anthropic==0.45.0
ollama==0.4.4
python-dotenv==1.0.1
